+++ src/compas_timber/connections/plate_butt_joint.py
@@ -8,9 +8,6 @@ from .plate_joint import move_polyline_segment_to_plane
 class PlateButtJoint(PlateJoint):
     """Creates a plate-to-plate butt-joint connection."""
 
-    def __init__(self, main_plate=None, cross_plate=None, **kwargs):
-        super(PlateButtJoint, self).__init__(main_plate, cross_plate, **kwargs)
-
     @property
     def main_plate(self):
         """Return the main plate."""