+++ src/compas_timber/model/model.py
@@ -13,6 +13,7 @@ from compas.tolerance import TOL
 from compas_model.models import Model
 
 from compas_timber.connections import ConnectionSolver
+from compas_timber.connections import Joint
 from compas_timber.connections import JointCandidate
 from compas_timber.connections import PlateJointCandidate
 from compas_timber.connections import Joint
@@ -398,7 +399,9 @@ class TimberModel(Model):
         for pair in pairs:
             beam_a, beam_b = pair
             result = solver.find_topology(beam_a, beam_b, max_distance=max_distance)
-            topology, beam_a, beam_b, distance, pt = result
+            topology = result.topology
+            beam_a = result.beam_a
+            beam_b = result.beam_b
             if topology == JointTopology.TOPO_UNKNOWN:
                 continue