+++ CHANGELOG.md
@@ -56,6 +56,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `JointTopology.TOPO_EDGE_EDGE` for Plate Connections.
 * Added `JointTopology.TOPO_EDGE_FACE` for Plate Connections.
 * Added `Cluster.topology`.
+* Added `PlateSolverResult` and `BeamSolverResult` to package results from `PlateConnectionSolver.find_topology()` and `ConnectionSolver.find_topology()`.
+* Added `PlateJointCandidate` to `generic_joint.py`.
+* Added `Joint.from_cluster` and `Joint.from_generic_joint` constructors to `Joint`.
+* Added `PlateJoint.from_generic_joint` as override.
 
 ### Changed
 
@@ -77,6 +81,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed default value for `modify_cross` to `True` for `LButtJoint`.
 * Minor fixes to GH Components.
 * Fixed `elements` and geometry creation for `BallNodeJoint`.
+* Changed `PlateConnectionSolver.find_topology()` to solve for `TOPO_EDGE_EDGE` or `TOPO_EDGE_FACE`.
+* Changed `PlateConnectionSolver.find_topology()` to return a `PlateSolverResult` instance.
+* Reworked `ConnectionSolver.find_topology()` for readability and to implement `TOPO_I`.
+* Changed `ConnectionSolver.find_topology()` to return a `BeamSolverResult` instance.
 * Removed `topology`, `a_segment_index` and `b_segment_index` from `PlateJoint` subclass `__init__()` methods. These can now be passed as kwargs.
 * `Platejoint`s can now be isntantiated with just 2 Plates as arguments. If no topology or segment index data is in kwargs, the joint will solve for those. 
 * Changed `PlateConnectionSolver.find_topology()` to return `TOPO_EDGE_EDGE` or `TOPO_EDGE_FACE`.